requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
ipykernel
numpy
pandas
matplotlib
seaborn 
flask
memory_profiler
streamlit
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
fitz
pymupdf
langchain-text-splitters
langchain-community
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
langserve
validators
ipykernel
numpy
pandas
matplotlib
seaborn 
flask
memory_profiler
streamlit
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
fitz
pymupdf
langchain-text-splitters
langchain-community
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
langserve
arxiv
wikipedia
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr